Fix compilation errors and 64-bit soundness in POSER
The last release of POSER was not built for anything newer than GCC
~3, and so would fail to compile with increased strictness of newer
compiler. Existing patches were written and provided on the dead
SourceForge project, and were applied directly to fix these known
compilation errors that affected the ability to compile POSER at
all.

Many areas of the POSER code assumed that `long` was a 32-bit type,
which is not true for 64-bit targets. These were just changed to
use `int32` or `uint32` as appropriate. Many areas also assumed that
it was possible and valid to cast any pointer type directly to/from
`uint32`. Most of the time, these are actually bogus pointers that
are smuggling `emuptr` values through the included Palm OS headers.
In one or two places, they tried to store host OS pointers inside
structures inside emulated memory, and that will never work. Those
areas are non-critical so just got a false assertion before the
invalid cast for now.

Other areas of the code used `sizeof` or `offsetof` on structs with
pointers in them, instead of the existing `EmAlias`/`EmProxy` system
which provides accurate information on the layout of structs in the
emulated memory. These invalid uses were replaced by `EmAlias`.

+++ PalmUtils/poser/SrcShared/ATraps.cpp
@@ -431,7 +431,7 @@ char* ATrap::GetStackBase ()
 {
 	// Ensure that the stack is aligned to a longword address.
 
-	uint32	stackBase = (uint32) fStack;
+	size_t	stackBase = (size_t) fStack;
 
 	stackBase += 3;
 	stackBase &= ~3;
